requirements.txt
discord.py>=2.4.0
aiosqlite>=0.20.0
APScheduler>=3.10.4
python-dotenv>=1.0.1
PyNaCl>=1.5.0
